requirements.txt
matplotlib==3.2.1
numpy==1.18.3
Pillow==7.1.2
scikit-learn==0.23.1
